backend/requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
psycopg[binary]==3.2.9
